refactor(scripts): extract platform-sync into pluggable adapter
- Extract GitHub sync checks (~700 lines) from validate-artifact.js
  into .agents/scripts/platform-adapters/platform-sync.js as an
  independent adapter loaded via top-level await at startup
- Rename verify.json key from "github-sync" to "platform-sync" across
  all 32 skill configs so switching platforms requires only swapping
  the adapter file with zero core or config changes
- Add graceful skip when no platform adapter is registered, with a
  regression test covering the missing-adapter scenario
