+++ .agents/skills/create-issue/config/verify.json
@@ -18,7 +18,7 @@
       "expected_action_pattern": "Create Issue",
       "freshness_minutes": 30
     },
-    "github-sync": {
+    "platform-sync": {
       "when": "issue_number_exists",
       "issue_must_exist": true,
       "verify_task_comment_content": true,
