+++ .agents/skills/create-task/config/verify.json
@@ -20,6 +20,6 @@
       "expected_action_pattern": "Task Created",
       "freshness_minutes": 30
     },
-    "github-sync": null
+    "platform-sync": null
   }
 }
