+++ .agents/skills/import-issue/config/verify.json
@@ -20,7 +20,7 @@
       "expected_action_pattern": "Import Issue",
       "freshness_minutes": 30
     },
-    "github-sync": {
+    "platform-sync": {
       "when": "issue_number_exists",
       "issue_must_exist": true
     }
