+++ .agents/skills/restore-task/config/verify.json
@@ -19,6 +19,6 @@
       "expected_action_pattern": "Restore Task",
       "freshness_minutes": 30
     },
-    "github-sync": null
+    "platform-sync": null
   }
 }
