+++ templates/.agents/skills/block-task/config/verify.json
@@ -20,7 +20,7 @@
       "expected_action_pattern": "Blocked",
       "freshness_minutes": 30
     },
-    "github-sync": {
+    "platform-sync": {
       "when": "issue_number_exists",
       "expected_status_label": "status: blocked"
     }
