+++ templates/.agents/skills/cancel-task/config/verify.json
@@ -21,7 +21,7 @@
       "expected_action_pattern": "Cancelled",
       "freshness_minutes": 30
     },
-    "github-sync": {
+    "platform-sync": {
       "when": "issue_number_exists",
       "expected_comment_marker": "<!-- sync-issue:{task-id}:cancel -->",
       "verify_task_comment_content": true
