+++ templates/.agents/skills/complete-task/config/verify.json
@@ -23,7 +23,7 @@
     "completion-checklist": {
       "require_all_checked": true
     },
-    "github-sync": {
+    "platform-sync": {
       "when": "issue_number_exists",
       "expected_comment_marker": "<!-- sync-issue:{task-id}:summary -->",
       "verify_task_comment_content": true,
