+++ templates/.agents/skills/refine-task/config/verify.json
@@ -25,7 +25,7 @@
       "expected_action_pattern": "Refinement \\(Round \\d+, for .+\\)",
       "freshness_minutes": 30
     },
-    "github-sync": {
+    "platform-sync": {
       "when": "issue_number_exists",
       "expected_status_label": "status: in-progress",
       "expected_comment_marker": "<!-- sync-issue:{task-id}:{artifact-stem} -->",
